# Integration of Nuget and Vsix into CI pipeline