--- a/.github/scripts/Env-Vars.ps1
+++ b/.github/scripts/Env-Vars.ps1
@@ -1,5 +1,0 @@
-# source this file to import these environment variables into a session
-$env:SAMPLES_VSIX_VERSION = "10.0.26100.0"
-$env:SAMPLES_VSIX_URI = "https://marketplace.visualstudio.com/_apis/public/gallery/publishers/DriverDeveloperKits-WDK/vsextensions/WDKVsix/10.0.26100.0/vspackage?targetPlatform=5e3e564c-03bb-4499-8ae5-b2b35e9a86dc"
-$env:SAMPLES_BUILD_NUMBER = "26100"
-$env:SAMPLES_NUGET_VERSION = "10.0.26100.1"