--- a/.github/workflows/ci-pr.yml
+++ b/.github/workflows/ci-pr.yml
@@ -22,8 +22,11 @@ jobs:
         with:
           submodules: 'recursive'
 
-      - name: Add MSBuild to PATH
-        uses: microsoft/setup-msbuild@v2
+      - name: Install WDK VSIX
+        run: .\.github\scripts\Install-Vsix.ps1
+
+      - name: Install Nuget Packages
+        run: nuget restore .\packages.config -PackagesDirectory .\packages\
 
       - name: Get changed files
         id: get-changed-files
@@ -63,8 +66,7 @@ jobs:
           path: _logs
 
       - name: Join and generate global reports
-        run: |
-          .\.github\scripts\Join-CsvReports.ps1
+        run: .\.github\scripts\Join-CsvReports.ps1
 
       - name: Archive global overview build reports
         uses: actions/upload-artifact@v3