… general/echo/kmdf/driver/AutoSync/echo.inx
… input/kbfiltr/sys/kbfiltr.inx
… sensors/Fusion/FusionSensor.inx