Adds Nuget package capability to the CI pipeline.

Adds installation of WDK vsix directly to the CI pipeline.

Updates environment variables for ease of incrementing component versions.

### Pipeline Time Metrics
**BuildAll** (push and pull_request)
Without Vsix Install: ~08m 59s
With Vsix Install:    ~38m 59s (optimized ~8m)

**BuildChange** (pull_request)
Without Vsix Install: ~03m 17s
With Vsix Install:    ~23m 20s (optimized "_pending_")

**CodeQL** (push and pull_request)
Without Vsix Install: ~1h 1m 43s
With Vsix Install:    ~1h 34m 31s (optimized ~1h 16m 2s)

**Average WDK Vsix Install Time**
~21m (optimized ~15s)

**Average WDK Nuget Restore**
~1m 30s

![image](https://github.com/microsoft/Windows-driver-samples/assets/138825652/e5d9b4a9-7dc5-49c2-a162-e946d593532f)

![image](https://github.com/microsoft/Windows-driver-samples/assets/138825652/023e3b54-8454-4a51-bfc4-69a6a92648e5)