@JakobL-MSFT 2024-06-05
Do you need to update your performance numbers now that you have $optimize?